Completed Udacity Review Recommendations
Even though the project was already complete, a Udacity grader
recommended I add the overall survival rates for the categories I chose.
I created a function that would complete this task. I also added a
function to print out stats about the age section so that it doesn't
seem so empty.